[#17] 로그인 뷰 UI 코드 이동
